requirements.txt
lxml==4.6.2
opcua==0.98.12
paho-mqtt==1.5.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
six==1.15.0
